requirements.txt
Django==2.0.2
django-tastypie==0.14.0
python-dateutil==2.6.1
python-mimeparse==1.6.0
six==1.11.0